4.21 done

+++ osa4/controllers/blogs.js
@@ -22,7 +22,6 @@ blogsRouter.post('/', async (request, response) => {
   
   const body = request.body
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  //const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
@@ -56,8 +55,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blogToBeDeleted = await Blog.findById(request.params.id)
+  if (blogToBeDeleted.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    response.status(400).end()
+  }
+
 })
 
 blogsRouter.post('/:id', (request, response) => {
